[#201] Fix import failure of utils and moved it
