+++ robot/basestation/stream_capture.py
@@ -5,7 +5,7 @@ import datetime
 import flask
 from flask import jsonify, make_response
 import threading
-from utils import run_shell
+from robot.util.utils import run_shell
 import robot.basestation.ros_utils as ros_utils
 from robot.basestation.ros_utils import fetch_ros_master_ip
